Move random helper functions to basetestclass
